move legacy ActivityFeed
