+++ src/components/legacy/ActivityFeed/ActivityFeed.js
@@ -32,7 +32,7 @@ import Title from 'components/Title';
 import Tabs from 'components/Tabs';
 import { BaseText, MediumText } from 'components/legacy/Typography';
 import EmptyStateParagraph from 'components/EmptyState/EmptyStateParagraph';
-import ActivityFeedItem from 'components/ActivityFeed/ActivityFeedItem';
+import ActivityFeedItem from 'components/legacy/ActivityFeed/ActivityFeedItem';
 import EventDetails, { shouldShowEventDetails } from 'components/EventDetails';
 import Modal from 'components/Modal';
