+++ src/screens/Collectible/Collectible.js
@@ -32,7 +32,7 @@ import { SEND_COLLECTIBLE_FROM_ASSET_FLOW } from 'constants/navigationConstants'
 import { COLLECTIBLE_TRANSACTION } from 'constants/collectiblesConstants';
 
 // components
-import ActivityFeed from 'components/ActivityFeed';
+import ActivityFeed from 'components/legacy/ActivityFeed';
 import ContainerWithHeader from 'components/legacy/Layout/ContainerWithHeader';
 import { ScrollWrapper, Wrapper } from 'components/legacy/Layout';
 import { Paragraph } from 'components/legacy/Typography';
