+++ src/screens/KeyBasedAssetTransfer/KeyBasedAssetTransferStatus.js
@@ -38,7 +38,7 @@ import {
 import ContainerWithHeader from 'components/legacy/Layout/ContainerWithHeader';
 import { Footer, ScrollWrapper, Wrapper } from 'components/legacy/Layout';
 import Button from 'components/legacy/Button';
-import ActivityFeed from 'components/ActivityFeed';
+import ActivityFeed from 'components/legacy/ActivityFeed';
 import Animation from 'components/Animation';
 import { MediumText, Paragraph } from 'components/legacy/Typography';
 import Spinner from 'components/Spinner';
